+++ gui/widgets/checkboxlist.cpp
@@ -1,21 +1,19 @@
 #include "checkboxlist.h"
 
 //min-width:10em;
-     CheckBoxList::CheckBoxList(QWidget *widget ) : QComboBox(widget),m_DisplayText("")
-     {
-    // set delegate items view
-    view()->setItemDelegate(new CheckBoxListDelegate(this));
-    //view()->setStyleSheet("  padding: 15px; ");
-    // Enable editing on items view
-    view()->setEditTriggers(QAbstractItemView::CurrentChanged);
-
-    // set "CheckBoxList::eventFilter" as event filter for items view
-    view()->viewport()->installEventFilter(this);
-
-
-    // it just cool to have it as defualt ;)
-    view()->setAlternatingRowColors(true);
-     }
+CheckBoxList::CheckBoxList(QWidget *widget ) : QComboBox(widget), m_DisplayText("")
+{
+	// set delegate items view
+	view()->setItemDelegate(new CheckBoxListDelegate(this));
+	//view()->setStyleSheet("  padding: 15px; ");
+	// Enable editing on items view
+	view()->setEditTriggers(QAbstractItemView::CurrentChanged | QAbstractItemView::DoubleClicked);
+	// set "CheckBoxList::eventFilter" as event filter for items view
+	view()->viewport()->installEventFilter(this);
+	// it just cool to have it as defualt ;)
+	view()->setAlternatingRowColors(true);
+	view()->setCurrentIndex(QModelIndex());
+}
 
 CheckBoxList::~CheckBoxList(){}
 
@@ -27,13 +25,17 @@ void CheckBoxList::hidePopup()
 
 bool CheckBoxList::eventFilter(QObject *object, QEvent *event)
 {
-    // don't close items view after we release the mouse button
-    // by simple eating MouseButtonRelease in viewport of items view
-    if(event->type() == QEvent::MouseButtonRelease && object==view()->viewport())
-    {
-        return true;
-    }
-    return QComboBox::eventFilter(object,event);
+	// Make sure the first event is not current, such that the editor opens on hover
+	if(event->type() == QEvent::Show && object==view()->viewport()) {
+		view()->setCurrentIndex(rootModelIndex());
+	}
+	// don't close items view after we release the mouse button
+	// by simple eating MouseButtonRelease in viewport of items view
+	if(event->type() == QEvent::MouseButtonRelease && object==view()->viewport())
+	{
+		return true;
+	}
+	return QComboBox::eventFilter(object,event);
 }
 
 void CheckBoxList::paintEvent(QPaintEvent *)
@@ -46,8 +48,8 @@ void CheckBoxList::paintEvent(QPaintEvent *)
     initStyleOption(&opt);
 
     // if no display text been set , use "..." as default
-    if(m_DisplayText.isNull())
-        opt.currentText = "......";
+    if(m_DisplayText.isEmpty())
+        opt.currentText = "....";
     else
         opt.currentText = m_DisplayText;
     painter.drawComplexControl(QStyle::CC_ComboBox, opt);
@@ -65,11 +67,11 @@ void CheckBoxList::setData(QStringList data)
 	for(int y = 0; y < model()->rowCount(); ++y)
     {
 		auto idx = model()->index(y,0);
-		if(data.contains(model()->data(idx,Qt::DisplayRole).toString()))
-		{
+		if(data.contains(model()->data(idx,Qt::DisplayRole).toString())) {
 			model()->setData(idx, true,Qt::UserRole);
 			data.removeAll(model()->data(idx,Qt::DisplayRole).toString());
-		}else{
+		}
+		else {
 			model()->setData(idx, false,Qt::UserRole);
 		}
     }
@@ -83,13 +85,12 @@ void CheckBoxList::setData(QStringList data)
 void CheckBoxList::updateText()
 {
 	QStringList value;
-    for(int y = 0; y < model()->rowCount(); ++y)
-    {
-        if(model()->data(model()->index(y,0),Qt::UserRole).toBool())
-        {
-            value.append(model()->data(model()->index(y,0),Qt::DisplayRole).toString());
-        }
-    }
+	for(int y = 0; y < model()->rowCount(); ++y)
+	{
+		if(model()->data(model()->index(y,0),Qt::UserRole).toBool()) {
+			value.append(model()->data(model()->index(y,0),Qt::DisplayRole).toString());
+		}
+	}
     
 	if (_current == value) return;
 	
@@ -98,25 +99,30 @@ void CheckBoxList::updateText()
 	emit currentTextChanged(_current);
 }
 
-QString	CheckBoxList::currentText() const
+QString CheckBoxList::currentText() const
 {
     return _current.join(", ");
 }
 
 QVariant CheckBoxList::currentData(int role) const
 {
-    if(role == Qt::UserRole || role == Qt::EditRole)
-    {
-        return _current;
-    }else if(role == Qt::DisplayRole){
+	if(role == Qt::UserRole || role == Qt::EditRole) {
+		return _current;
+	}
+	else if(role == Qt::DisplayRole){
 		return m_DisplayText;
-	}else{
-        return false;
-    }
+	}
+	else{
+		return false;
+	}
 }
+
+
+CheckBoxListDelegate::CheckBoxListDelegate(QObject* parent) : QAbstractItemDelegate(parent) {}
 
 void CheckBoxListDelegate::paint(QPainter *painter, const QStyleOptionViewItem &option,const QModelIndex &index) const
 {
+	// Draw CheckBox views as visualisation
     bool value = index.data(Qt::UserRole).toBool();
     QString text = index.data(Qt::DisplayRole).toString();
     const QStyle *style = QApplication::style();
@@ -135,18 +141,23 @@ QSize CheckBoxListDelegate::sizeHint(const QStyleOptionViewItem & option, const 
     return QSize(90,30);
 }
 
+bool CheckBoxListDelegate::editorEvent(QEvent *event, QAbstractItemModel *model, const QStyleOptionViewItem &option, const QModelIndex &index) {
+	return false;
+}
+
 QWidget* CheckBoxListDelegate::createEditor(QWidget *parent,const QStyleOptionViewItem & option ,const QModelIndex & index ) const
 {
-    QCheckBox* edit = new QCheckBox(parent);
-    // this->connect(edit,SIGNAL(stateChanged()),this,SLOT(setModelData()));
-    connect(edit, &QPushButton::clicked, [=] {setModelData(edit,const_cast<QAbstractItemModel*>(index.model()),index);});
-    return edit;
+	QCheckBox* edit = new QCheckBox(parent);
+	edit->setFocusPolicy(Qt::StrongFocus);
+	// this->connect(edit,SIGNAL(stateChanged()),this,SLOT(setModelData()));
+	connect(edit, &QPushButton::clicked, [=] {setModelData(edit,const_cast<QAbstractItemModel*>(index.model()),index);});
+	return edit;
 }
 
 void CheckBoxListDelegate::setEditorData(QWidget *editor, const QModelIndex &index) const
 {
     QCheckBox* edit = static_cast<QCheckBox*>(editor);
-    edit->setText(index.data(Qt::DisplayRole).toString());
+//     edit->setText(index.data(Qt::DisplayRole).toString());
     edit->setChecked(index.data(Qt::UserRole).toBool());
 }
 
@@ -157,7 +168,7 @@ void CheckBoxListDelegate::setModelData(QWidget *editor, QAbstractItemModel *mod
 
     //set model data
     QMap<int,QVariant> data;
-    data.insert(Qt::DisplayRole,myEditor->text());
+//     data.insert(Qt::DisplayRole,myEditor->text());
     data.insert(Qt::UserRole,value);
     model->setItemData(index,data);
 }